[Bronze V] Title: 이상한 기호, Time: 68 ms, Memory: 11516 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,12 +5,13 @@ class Main {
     public static void main(String[] args) throws Exception {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int A = Integer.parseInt(st.nextToken());
-        int B = Integer.parseInt(st.nextToken());
-        System.out.println(culAB(A,B));
+        long A = Long.parseLong(st.nextToken());
+        long B = Long.parseLong(st.nextToken());
+
+        System.out.println(culAB(A, B));
     }
 
-    public static int culAB(int A, int B) {
-        return (A+B)*(A-B);
+    public static long culAB(long A, long B) {
+        return (A + B) * (A - B);
     }
 }
